+++ src/main/java/ru/dnlkk/ratingusbackend/api/ProfileApi.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import ru.dnlkk.ratingusbackend.api.model.UserDto;
+import ru.dnlkk.ratingusbackend.api.dtos.UserDto;
 
 
 @Tag(name = "Контроллер профиля", description = "Просмотр и редактирование профиля")
