+++ src/main/java/ru/dnlkk/ratingusbackend/api/dtos/AnnouncementDto.java
@@ -1,4 +1,4 @@
-package ru.dnlkk.ratingusbackend.api.model;
+package ru.dnlkk.ratingusbackend.api.dtos;
 
 import lombok.*;
