--- src/main/java/ru/dnlkk/ratingusbackend/api/model/LessonDto.java
@@ -1,4 +1,0 @@
-package ru.dnlkk.ratingusbackend.api.model;
-
-public class LessonDto {
-}
