+++ src/main/java/ru/dnlkk/ratingusbackend/model/Subject.java
@@ -1,4 +1,4 @@
-package ru.dnlkk.ratingusbackend.dtos;
+package ru.dnlkk.ratingusbackend.model;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
@@ -6,31 +6,30 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.sql.Timestamp;
+import java.util.List;
 
 @Setter
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
 @Entity
-@Table(name = "timetable")
-public class Timetable {
+@Table(name = "studies")
+public class Subject {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int id;
 
-    @Column(name = "lesson_number")
-    private int lessonNumber;
+    @Column(name = "name")
+    private String name;
 
-    @Column(name = "start_time")
-    private Timestamp startTime;
+    @ManyToMany(mappedBy = "subjects")
+    private List<User> teachers;
 
-    @Column(name = "end_time")
-    private Timestamp endTime;
+    @OneToMany(mappedBy = "subject")
+    private List<Lesson> lessons;
 
     @ManyToOne
     @JoinColumn(name = "school_id", referencedColumnName = "id")
     private School school;
-
 }
