+++ src/main/java/ru/dnlkk/ratingusbackend/model/UserCode.java
@@ -1,4 +1,4 @@
-package ru.dnlkk.ratingusbackend.dtos;
+package ru.dnlkk.ratingusbackend.model;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
