+++ src/main/java/ru/dnlkk/ratingusbackend/repository/AnnouncementRepository.java
@@ -3,7 +3,7 @@ package ru.dnlkk.ratingusbackend.repository;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import ru.dnlkk.ratingusbackend.dtos.Announcement;
+import ru.dnlkk.ratingusbackend.model.Announcement;
 
 import java.util.List;
